refactor(doctor): extract mutable allowlist warning helpers
