+++ src/commands/doctor-config-flow.ts
@@ -12,7 +12,6 @@ import { getChannelsCommandSecretTargetIds } from "../cli/command-secret-targets
 import { listRouteBindings } from "../config/bindings.js";
 import type { OpenClawConfig } from "../config/config.js";
 import { CONFIG_PATH, migrateLegacyConfig, readConfigFileSnapshot } from "../config/config.js";
-import { collectProviderDangerousNameMatchingScopes } from "../config/dangerous-name-matching.js";
 import { formatConfigIssueLines } from "../config/issue-format.js";
 import { applyPluginAutoEnable } from "../config/plugin-auto-enable.js";
 import { parseToolsBySenderTypedKey } from "../config/types.tools.js";
@@ -37,14 +36,6 @@ import {
   normalizeAccountId,
   normalizeOptionalAccountId,
 } from "../routing/session-key.js";
-import {
-  isDiscordMutableAllowEntry,
-  isGoogleChatMutableAllowEntry,
-  isIrcMutableAllowEntry,
-  isMSTeamsMutableAllowEntry,
-  isMattermostMutableAllowEntry,
-  isSlackMutableAllowEntry,
-} from "../security/mutable-allowlist-detectors.js";
 import { inspectTelegramAccount } from "../telegram/account-inspect.js";
 import { listTelegramAccountIds, resolveTelegramAccount } from "../telegram/accounts.js";
 import { note } from "../terminal/note.js";
@@ -57,6 +48,7 @@ import {
   stripUnknownConfigKeys,
 } from "./doctor-config-analysis.js";
 import { normalizeCompatibilityConfigValues } from "./doctor-legacy-config.js";
+import { scanMutableAllowlistEntries } from "./doctor-mutable-allowlist-warnings.js";
 import type { DoctorOptions } from "./doctor-prompter.js";
 import { autoMigrateLegacyStateDir } from "./doctor-state-migrations.js";
 
@@ -614,278 +606,6 @@ function maybeRepairDiscordNumericIds(cfg: OpenClawConfig): {
     return { config: cfg, changes: [] };
   }
   return { config: next, changes };
-}
-
-type MutableAllowlistHit = {
-  channel: string;
-  path: string;
-  entry: string;
-  dangerousFlagPath: string;
-};
-
-function addMutableAllowlistHits(params: {
-  hits: MutableAllowlistHit[];
-  pathLabel: string;
-  list: unknown;
-  detector: (entry: string) => boolean;
-  channel: string;
-  dangerousFlagPath: string;
-}) {
-  if (!Array.isArray(params.list)) {
-    return;
-  }
-  for (const entry of params.list) {
-    const text = String(entry).trim();
-    if (!text || text === "*") {
-      continue;
-    }
-    if (!params.detector(text)) {
-      continue;
-    }
-    params.hits.push({
-      channel: params.channel,
-      path: params.pathLabel,
-      entry: text,
-      dangerousFlagPath: params.dangerousFlagPath,
-    });
-  }
-}
-
-function scanMutableAllowlistEntries(cfg: OpenClawConfig): MutableAllowlistHit[] {
-  const hits: MutableAllowlistHit[] = [];
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "discord")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.allowFrom`,
-      list: scope.account.allowFrom,
-      detector: isDiscordMutableAllowEntry,
-      channel: "discord",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    const dm = asObjectRecord(scope.account.dm);
-    if (dm) {
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.dm.allowFrom`,
-        list: dm.allowFrom,
-        detector: isDiscordMutableAllowEntry,
-        channel: "discord",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-    const guilds = asObjectRecord(scope.account.guilds);
-    if (!guilds) {
-      continue;
-    }
-    for (const [guildId, guildRaw] of Object.entries(guilds)) {
-      const guild = asObjectRecord(guildRaw);
-      if (!guild) {
-        continue;
-      }
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.guilds.${guildId}.users`,
-        list: guild.users,
-        detector: isDiscordMutableAllowEntry,
-        channel: "discord",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-      const channels = asObjectRecord(guild.channels);
-      if (!channels) {
-        continue;
-      }
-      for (const [channelId, channelRaw] of Object.entries(channels)) {
-        const channel = asObjectRecord(channelRaw);
-        if (!channel) {
-          continue;
-        }
-        addMutableAllowlistHits({
-          hits,
-          pathLabel: `${scope.prefix}.guilds.${guildId}.channels.${channelId}.users`,
-          list: channel.users,
-          detector: isDiscordMutableAllowEntry,
-          channel: "discord",
-          dangerousFlagPath: scope.dangerousFlagPath,
-        });
-      }
-    }
-  }
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "slack")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.allowFrom`,
-      list: scope.account.allowFrom,
-      detector: isSlackMutableAllowEntry,
-      channel: "slack",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    const dm = asObjectRecord(scope.account.dm);
-    if (dm) {
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.dm.allowFrom`,
-        list: dm.allowFrom,
-        detector: isSlackMutableAllowEntry,
-        channel: "slack",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-    const channels = asObjectRecord(scope.account.channels);
-    if (!channels) {
-      continue;
-    }
-    for (const [channelKey, channelRaw] of Object.entries(channels)) {
-      const channel = asObjectRecord(channelRaw);
-      if (!channel) {
-        continue;
-      }
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.channels.${channelKey}.users`,
-        list: channel.users,
-        detector: isSlackMutableAllowEntry,
-        channel: "slack",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-  }
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "googlechat")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.groupAllowFrom`,
-      list: scope.account.groupAllowFrom,
-      detector: isGoogleChatMutableAllowEntry,
-      channel: "googlechat",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    const dm = asObjectRecord(scope.account.dm);
-    if (dm) {
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.dm.allowFrom`,
-        list: dm.allowFrom,
-        detector: isGoogleChatMutableAllowEntry,
-        channel: "googlechat",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-    const groups = asObjectRecord(scope.account.groups);
-    if (!groups) {
-      continue;
-    }
-    for (const [groupKey, groupRaw] of Object.entries(groups)) {
-      const group = asObjectRecord(groupRaw);
-      if (!group) {
-        continue;
-      }
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.groups.${groupKey}.users`,
-        list: group.users,
-        detector: isGoogleChatMutableAllowEntry,
-        channel: "googlechat",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-  }
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "msteams")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.allowFrom`,
-      list: scope.account.allowFrom,
-      detector: isMSTeamsMutableAllowEntry,
-      channel: "msteams",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.groupAllowFrom`,
-      list: scope.account.groupAllowFrom,
-      detector: isMSTeamsMutableAllowEntry,
-      channel: "msteams",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-  }
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "mattermost")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.allowFrom`,
-      list: scope.account.allowFrom,
-      detector: isMattermostMutableAllowEntry,
-      channel: "mattermost",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.groupAllowFrom`,
-      list: scope.account.groupAllowFrom,
-      detector: isMattermostMutableAllowEntry,
-      channel: "mattermost",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-  }
-
-  for (const scope of collectProviderDangerousNameMatchingScopes(cfg, "irc")) {
-    if (scope.dangerousNameMatchingEnabled) {
-      continue;
-    }
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.allowFrom`,
-      list: scope.account.allowFrom,
-      detector: isIrcMutableAllowEntry,
-      channel: "irc",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    addMutableAllowlistHits({
-      hits,
-      pathLabel: `${scope.prefix}.groupAllowFrom`,
-      list: scope.account.groupAllowFrom,
-      detector: isIrcMutableAllowEntry,
-      channel: "irc",
-      dangerousFlagPath: scope.dangerousFlagPath,
-    });
-    const groups = asObjectRecord(scope.account.groups);
-    if (!groups) {
-      continue;
-    }
-    for (const [groupKey, groupRaw] of Object.entries(groups)) {
-      const group = asObjectRecord(groupRaw);
-      if (!group) {
-        continue;
-      }
-      addMutableAllowlistHits({
-        hits,
-        pathLabel: `${scope.prefix}.groups.${groupKey}.allowFrom`,
-        list: group.allowFrom,
-        detector: isIrcMutableAllowEntry,
-        channel: "irc",
-        dangerousFlagPath: scope.dangerousFlagPath,
-      });
-    }
-  }
-
-  return hits;
 }
 
 /**
